Make new dag_run for each run of airflow-cwl-runner
Each run of airflow-cwl-runner creates new record in dag_run table.
If we run it with the same CWL descriptor and input parameters
files dag_id, which is generated on the base if this filenames, will
be also the same. In this case backfill won't run the tasks and
airflow-cwl-runner will return the results of the prevously
calculated tasks. To avoid it we check if we already have dag runs
with the same dag_id in dag_run table. If yes, we add the sufix _N
to the end of dag_id