Search for cwl jobs in subfolders 'new' in CWL_JOBS
When setting folder to search for input parameters files for running
workflow it's better to search for any directory/subdirectory with
the name 'new', after that check if there are 'running', 'fail',
'success' folder on the same level as 'new' one. If some of them
are not found, raise an error. It allows to make separate folders
for job files with different access permissions