Update setting JobDispatcher as upstream task
Previously JobDispatcher task was set as upstream
for all tools that have inputs with 'source' in
workflow inputs. All tool inputs with 'default'
were additionally excluded to prevent setting
JobDispatcher as an upstream for those tools,
which are run from workflow with default parameters.
That's mistake.
To fix it - we exclude inputs only if they don't have
'source'. It means that default is set on the workflow
level. If tool have both 'default' and 'source' it means
default is set on tool level, but we still need
to connect it to JobDispatcher upstream task