Fix bug in setting inputs for tool
When setting input values for tool
we check if 'source' from tool input
corresponds to 'id' from workflow input.
If we didn't find such correspondence
we skip this tool input. But in a case
when tool input has 'source' set to
optional workflow input and then
depending on whether it found it or not
it performs some calculation in valueFrom,
there is a chance that we skip this tool
input by mistake. To fix it, instead of
skipping tool input, we set it to None.
In this case if there is valueFrom in
tool input, it will be calculated. If not
the default value from tool itself will
be used.