Update the way of choosing the job for execution
When we scan CWL_JOBS folder for all new job
files in each of the 'new' subfolders, there
is a chance that we accidentally rewrite job
file in the 'running' subfolder with the same
name. To prevent it, we first filter out all
job files in all 'new' folders that have the
same name as job files in correspondent
'running' subfolders. And only after that we
sort all these files by last update time and
choose the oldest one to be moved to
correspondent 'running' subfolder.