Fix bug of setting Null instead of Default
When setting values for step inputs, we
iterate over all data received from the
previous step and if we find 'id' which
corresponds to input 'source', we set
it to this value. If some input has
'source' field, but we didn't manage to
find correspondent field in upstream data
with 'id'='source', we should skip this
input. But then we cannot calculate
'valueFrom'. If don't skip this input and
set it to Null it also causes a problem in
a case if this input has Default value in
embedded tool. If it's set to Null cwltool
use Null, but not Default (which is not so
logical, imho). The solution is to set Null
only in that case, when 'source'!='id' and
'valueFrom' exists. If we couldn't find
even 'valueFrom', we definately should skip
this input
ADDITIONALLY: when using airflow-cwl-runner,
generate unique dag_id (check for existing
dag_id's in DB)