Support running cwl#fragment
When workflow is packed with --print-pre
to run it the fragment should be specified
in a form of workflow_name.cwl#main, where
this fragment "main" will be added to all
step "source" and "id". This means that
jobdispatcher task should generate job
object with this prefix too. To allow
this we check if workflow argument includes
some fragment, and if yes, add it before
all "id" in generated job object.